More output in realtime binning

+++ deepbinner/realtime.py
@@ -17,6 +17,7 @@ import pathlib
 import shutil
 import sys
 import time
+from .misc import print_summary_table
 
 from .classify import load_and_check_models, classify_fast5_files, set_tensorflow_threads
 from .bin import class_to_class_names
@@ -38,8 +39,6 @@ def realtime(args):
         while True:
             fast5s = look_for_new_fast5s(args.in_dir, args.out_dir, nested_out_dir)
             if fast5s:
-                if waiting:
-                    print()
                 time.sleep(5)  # wait a bit to make sure any file moves are finished
                 classify_and_move(fast5s, args, start_model, start_input_size, end_model,
                                   end_input_size, output_size)
@@ -71,21 +70,48 @@ def classify_and_move(fast5s, args, start_model, start_input_size, end_model, en
     classifications, read_id_to_fast5_file = \
         classify_fast5_files(fast5s, start_model, start_input_size, end_model, end_input_size,
                              output_size, args, full_output=False)
+    print()
+
+    move_count, fail_move_already_exists, fail_move_other_reason = 0, 0, 0
     counts = collections.defaultdict(int)
     for read_id, barcode_call in classifications.items():
         fast5_file = read_id_to_fast5_file[read_id]
+
         out_dir = pathlib.Path(args.out_dir) / class_to_class_names(barcode_call)
         if not out_dir.is_dir():
             try:
                 os.makedirs(str(out_dir))
             except (FileNotFoundError, OSError, PermissionError):
                 sys.exit('Error: unable to create output directory {}'.format(out_dir))
-        try:
-            shutil.move(fast5_file, out_dir)
-        except (FileNotFoundError, OSError, PermissionError):
-            print('Error: could not move {} to {}'.format(fast5_file, out_dir))
+
+        dest_filepath = out_dir / pathlib.Path(fast5_file).name
+        if dest_filepath.is_file():
+            fail_move_already_exists += 1
+        else:
+            try:
+                shutil.move(fast5_file, out_dir)
+            except (FileNotFoundError, OSError, PermissionError):
+                fail_move_other_reason += 1
+
         counts[barcode_call] += 1
+        move_count += 1
+        print_moving_progress(move_count, len(fast5s))
+
     print()
+    print_moving_error_messages(fail_move_already_exists, fail_move_other_reason, args.out_dir)
+    print_summary_table(classifications, output=sys.stdout)
+
+
+def print_moving_error_messages(already_exists, other_reason, out_dir):
+    if already_exists == 1:
+        print('Error: could not move 1 fast5 file because it already exists in {}'.format(out_dir))
+    elif already_exists > 1:
+        print('Error: could not move {} fast5 files because they already exist '
+              'in {}'.format(already_exists, out_dir))
+    if other_reason == 1:
+        print('Error: failed to move 1 fast5 file to {}'.format(out_dir))
+    elif other_reason > 1:
+        print('Error: failed to move {} fast5 files to {}'.format(other_reason, out_dir))
 
 
 def make_output_dir(out_dir):
@@ -98,3 +124,9 @@ def make_output_dir(out_dir):
             print('Making output directory: {}/'.format(out_dir))
         except (FileNotFoundError, OSError, PermissionError):
             sys.exit('Error: unable to create output directory {}'.format(out_dir))
+
+
+def print_moving_progress(completed, total):
+    percent = 100.0 * completed / total
+    print('\rMoving fast5s:      {} / {} ({:.1f}%)'.format(completed, total, percent),
+          end='', flush=True)
